+++ src/Objects/GameObjects/GameObjectItem/Teleport.cpp
@@ -8,5 +8,5 @@ void Teleport::load(const int x, const int y, const int width, const int height,
     destMap = pCustomProperties.getString("destMap","Missing map name");
     destX = pCustomProperties.getFloat("destX");
     destY = pCustomProperties.getFloat("destY");
-    destDirection = static_cast<DIRECTION>(pCustomProperties.getInt("destDirection", DIRECTION::SOUTH));
+    destDirection = static_cast<DIRECTION>(pCustomProperties.getInt("destDirection", static_cast<int>(DIRECTION::SOUTH)));
 }
